Componentes del formulario, endpoint+fetch userID Ahora hay dos archivos que representan dos páginas distintas: 1) CreateNewUserProfile, 2) EditUserProfile Dentro de esos archivos ahora solo hay componentes (dos copias del formulario pero con variaciones).
He creado un nuevo endpoint para conseguir el id_user usando su correo.
En flux he creado un nuevo fetch POST para ese endpoint.

Ahora después de iniciar sesión el id del usuario en la table de Users
se guarda en una variable dentro del store (he añadido una función de la página de Signup).
Podemos usar ese id para hacer llamadas Fetch GET a la tabla de usuarios y pintar sus datos en la página.
