+++ src/front/js/layout.js
@@ -16,6 +16,7 @@ import { EventSearchJoin } from "./pages/EventSearchJoin";
 import EventPagePay from "./pages/EventPagePay";
 
 import { CreateNewUserProfile } from "./pages/CreateNewUserProfile";
+import { EditUserProfile } from "./pages/EditUserProfile"
 
 
 // Components
@@ -47,6 +48,8 @@ const Layout = () => {
                         <Route element={<EventPagePay />} path="/EventPagePay/:id" />
 
                         <Route element={<CreateNewUserProfile />} path="/CreateNewUserProfile" />
+                        <Route element={<EditUserProfile />} path="/EditUserProfile" />
+
                         <Route element={<Single />} path="/single/:theid" />
                         <Route element={<h1>Not found!</h1>} />
                     </Routes>
