+++ src/front/js/pages/EditUserProfile.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { FormEditUser } from "../component/FormEditUser"
+
+export const EditUserProfile = () => {
+  return (
+          <FormEditUser/>
+  );
+};
