+++ src/front/js/pages/SignIn.js
@@ -22,6 +22,7 @@ export const SignIn = () => {
 
         // CONDITIONALS with FRONTEND messages (below in the jsx):
         setTimeout(() => {
+            actions.getUserID(email);
             actions.setRegistrationInProgress(false);
             if (store.registrationSuccess) {
                 actions.setRegistrationSuccess(false);
