+++ src/front/styles/createNewUserProfile.css
@@ -27,7 +27,7 @@
    background-color: #018fa6;
 }
 
-.form-control:focus {
+.form-control:focus, .form-select:focus {
     border-color: #018fa6;
     box-shadow: 0px 1px 1px rgba(0, 0, 0, 0.075) inset, 0px 0px 12px #018fa6;
 }
